requirements.txt
# API Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Base de données
sqlmodel==0.0.24
psycopg2-binary==2.9.9  # Pour PostgreSQL
alembic==1.13.3  # Migrations

# Authentification
bcrypt==4.3.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
authlib==1.3.2  # Pour OAuth2

# RSS et parsing
feedparser==6.0.11
bleach==6.2.0
apscheduler==3.10.4  # Pour les tâches en arrière-plan

# Utilitaires
python-dateutil==2.8.2
requests==2.32.3
httpx==0.27.2  # Requis par authlib
itsdangerous==2.2.0  # Pour les sessions
python-dotenv==1.1.1  # Variables d'environnement

# Envoi d'emails pour 2FA
yagmail==0.15.293  # Envoi d'emails simplifié
secure-smtplib==0.1.1  # SMTP sécurisé